pocket-signals-bot/requirements.txt
aiogram==3.10.0
playwright==1.45.0
python-dotenv==1.0.1
asyncio==3.4.3
websockets==12.0
